+++ lib/widgets/pixabay/by_pixabay.dart
@@ -27,7 +27,7 @@ class ByPixabay extends StatelessWidget {
             ),
             MouseRegion(
               cursor: SystemMouseCursors.click,
-              child: TextButton(
+              child: HyperlinkButton(
                   style: ButtonStyle(
                       padding: ButtonState.all(const EdgeInsets.all(0))),
                   child: const Text('Pixabay'),