+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
+## 1.7.0
+
+- update to flutter v3.0
+- update dependencies versions
+
 ## 1.5.0
 
-- option to set  watch background from [Pixabay](https://pixabay.com) images
+- option to set watch background from [Pixabay](https://pixabay.com) images
 - option to set desktop-background (wallpaper) from [Pixabay](https://pixabay.com) images
 - bug fixes and improvements
 
 ## 1.2.0
 
-- enable to select a local background image 
+- enable to select a local background image